refactored pools.ts to use only one pool config file
